+++ src/ui/popover/wasalert/wasalert.dialog.ts
@@ -60,8 +60,11 @@ export class WasAlert {
     public dialogRef: MatDialogRef<WasAlert>,
     @Inject(MAT_DIALOG_DATA) public data: any) {
 
+    // console.log('print wasalert ', dialogRef);
+
     // SET DEFAULT VALUES
     dialogRef.disableClose = true; // do not close by clicking off by default
+    if (!this.data) { this.data = {}; } // data may not be defined
     if (!this.data.title) { this.data.title = ''; }
     if (!this.data.body) { this.data.body = ''; }
     if (!this.data.buttons) { this.data.buttons = ['Cancel']; }