+++ CHANGELOG.md
@@ -9,10 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 This shows all commits since last release (actually, currently all commits).
 
-## [2.1.0](https://bitbucket.org/wickeyware/wickeyappstore_npm/commits/tag/2.1.0) - 2018-21-03
+## [2.1.0](https://bitbucket.org/wickeyware/wickeyappstore_npm/commits/tag/2.1.0) - 2018-23-03
 
 ### Added
-
+* **WasPay** Add simple purchase interface.
+* **WasShop** Add InApp Shop interface.
 * **WasAppService** Add bannerApps service.
 * **WasAppService** Add featuredGroups service.
 * **API** Add API to get Inapps.